requirements.txt
# Python 3.10+
anthropic==0.40.0
networkx==3.4.2
chromadb==0.5.20
sentence-transformers==3.3.1
numpy==2.1.3
pyyaml==6.0.2
python-dotenv==1.0.1
aiohttp==3.11.11
beautifulsoup4==4.12.3
apscheduler==3.10.4
markdown==3.7
pdfplumber==0.11.4
FlagEmbedding==1.2.10
